Filters refactored (closes #52)
