--- src/PHPSC/Conference/Application/Action/NewAttendeeFilter.php
@@ -1,28 +1,0 @@
-<?php
-namespace PHPSC\Conference\Application\Action;
-
-class NewAttendeeFilter extends BasicFilter
-{
-    /**
-     * @see \Lcobucci\ActionMapper2\Routing\Filter::process()
-     */
-    public function process()
-    {
-        $this->validateUserRegistration();
-
-        $event = $this->getEventManagement()->findCurrentEvent();
-
-        if (!$event->isRegistrationInterval(new \DateTime())) {
-            $this->application->redirect('/registration');
-        }
-    }
-
-    /**
-     * @return \PHPSC\Conference\Domain\Service\EventManagementService
-     */
-    protected function getEventManagement()
-    {
-        return $this->application->getDependencyContainer()
-                                 ->get('event.management.service');
-    }
-}
