+++ src/PHPSC/Conference/Application/Filter/NewAttendeeFilter.php
@@ -1,0 +1,30 @@
+<?php
+namespace PHPSC\Conference\Application\Filter;
+
+use DateTime;
+use PHPSC\Conference\Domain\Service\EventManagementService;
+
+class NewAttendeeFilter extends BasicFilter
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function process()
+    {
+        $this->validateUserRegistration();
+
+        $event = $this->getEventManagement()->findCurrentEvent();
+
+        if (!$event->isRegistrationInterval(new DateTime())) {
+            $this->application->redirect('/registration');
+        }
+    }
+
+    /**
+     * @return EventManagementService
+     */
+    protected function getEventManagement()
+    {
+        return $this->get('event.management.service');
+    }
+}
